requirements.txt
lxml==4.6.2
requests==2.25.1
beautifulsoup4==4.9.3
Pillow==8.1.2
